# openapi.yaml
openapi: 3.0.3
info:
  title: Advertisement Service
  version: 1.0.0
tags:
  - name: advertisement
  - name: user
paths:
  /advertisement:
    post:
      tags:
        - advertisement
      summary: Add new advertisement
      operationId: addAdvertisement
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewAdvertisement'
          application/xml:
            schema:
              $ref: '#/components/schemas/NewAdvertisement'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewAdvertisement'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Advertisement'
            application/xml:
              schema:
                $ref: '#/components/schemas/Advertisement'
        '405':
          description: Invalid input
      security:
        - api_key: []
  /advertisement/findByCategory:
    get:
      tags:
        - advertisement
      summary: Find advertisements by category
      operationId: findAdvertisementBySategory
      parameters:
        - name: category
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Advertisement'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Advertisement'
        '400':
          description: Invalid category value
  /advertisement/findByAccess:
    get:
      tags:
        - advertisement
      summary: Find advertisements by access
      operationId: findAdvertisementByAccess
      parameters:
        - name: access
          in: query
          required: true
          schema:
            type: string
            enum:
              - public
              - local
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Advertisement'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Advertisement'
        '400':
          description: Invalid access value
        '401':
          description: Unauthorized
        '403':
          description: Access denied
  /advertisement/findByUser:
    get:
      tags:
        - advertisement
      summary: Find advertisement by username
      operationId: findAdvertisementByUsername
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Advertisement'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Advertisement'
        '400':
          description: Ads doesn't exists
      security:
        - api_key: []
  /advertisement/:
    get:
      tags:
        - advertisement
      summary: Find advertisement by ID
      operationId: getAdvertisementById
      parameters:
        - name: advertisementId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Advertisement'
            application/xml:
              schema:
                $ref: '#/components/schemas/Advertisement'
        '400':
          description: Invalid ID supplied
        '403':
          description: Access denied

      security:
        - api_key: []
    put:
      tags:
        - advertisement
      summary: Update advertisement by id
      operationId: updateAdvertisementWithForm
      parameters:
        - name: advertisementId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Advertisement'
          application/xml:
            schema:
              $ref: '#/components/schemas/Advertisement'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Advertisement'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Advertisement'
            application/xml:
              schema:
                $ref: '#/components/schemas/Advertisement'
        '400':
          description: Invalid ID supplied
        '403':
          description: Access denied
        '404':
          description: Advertisement not found
        '405':
          description: Invalid input
      security:
        - api_key: []
    delete:
      tags:
        - advertisement
      summary: Deletes a advertisement
      operationId: deleteAdvertisement
      parameters:
        - name: advertisementId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Advertisement'
            application/xml:
              schema:
                $ref: '#/components/schemas/Advertisement'
        '400':
          description: Invalid ID supplied
        '403':
          description: Access denied
        '404':
          description: Advertisement not found
      security:
        - api_key: []
  /category/register:
    post:
      tags:
        - category
      summary: Registrate
      operationId: registrateCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
          application/xml:
            schema:
              $ref: '#/components/schemas/Category'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Category'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
            application/xml:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Invalid input
        '401':
          description: Not logged in
      security:
        - api_key: []
  /category:
    delete:
      tags:
        - category
      summary: Delete category
      operationId: deleteCategory
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
            application/xml:
              schema:
                $ref: '#/components/schemas/Category'

        '401':
          description: Not logged in
        '405':
          description: Category doesn't exists
      security:
          - api_key: []
    get:
      tags:
        - category
      summary: Find advertisement by ID
      operationId: getCategoryById
      parameters:
        - name: categorytId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
            application/xml:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Invalid ID supplied

        '405':
          description: Category not found

  /user/register:
    post:
      tags:
        - user
      summary: Registrate
      operationId: registrateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
          application/xml:
            schema:
              $ref: '#/components/schemas/NewUser'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewUser'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'

        '400':
          description: Not correct data provided

  /user/:
    get:
      tags:
        - user
      summary: Get user by user id
      description: ''
      operationId: getUserById
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid username supplied
        '401':
          description: Not logged in


        '405':
          description: User not found
      security:
        - api_key: []
  /user/username:
    get:
      tags:
        - user
      summary: Get user by user name
      description: ''
      operationId: getUserByName
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewUser'
            application/xml:
              schema:
                $ref: '#/components/schemas/NewUser'
        '400':
          description: Invalid username supplied
        '401':
          description: Not logged in
        '403':
          description: Access denied

        '405':
          description: User not found
      security:
        - api_key: []
    put:
      tags:
        - user
      summary: Update user
      operationId: updateUser
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent user in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          application/xml:
            schema:
              $ref: '#/components/schemas/User'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewUser'
            application/xml:
              schema:
                $ref: '#/components/schemas/NewUser'
        '400':
          description: Invalid username supplied
        '403':
          description: Access denied
        '405':
          description: Invalid input
      security:
        - api_key: []
    delete:
      tags:
        - user
      summary: Delete user
      operationId: deleteUser
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid username supplied
        '401':
          description: Incorrect password or username
        '403':
          description: Access denied
        '404':
          description: User not found
      security:
          - api_key: []
components:
  schemas:
    NewAdvertisement:
      type: object
      properties:
        description:
          type: string
          example: missing doggie
        is_global:
          type: string
          enum:
            - local
            - public
        category_id:
          $ref:  '#/components/schemas/Category'
        user_id:
          $ref:  '#/components/schemas/User'

      xml:
        name: advertisement
    Advertisement:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        description:
          type: string
          example: missing doggie

        is_global:
          type: string
          enum:
            - local
            - public
        category_id:
          $ref:  '#/components/schemas/Category'
        user_id:
          $ref:  '#/components/schemas/User'
      xml:
        name: advertisement

    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: missing pet
      xml:
        name: category
    User:

      type: object
      properties:
        id:
            type: integer
            format: int64
            example: 1
        username:
          type: string
          example: theUser
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: James
        address:
          type: string
          example: Street smth
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: '12345'
        phone_number:
          type: string
          example: '12345'
      xml:
        name: user
    NewUser:
      type: object
      properties:
        username:
          type: string
          example: theUser
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: James
        addres:
          type: string
          example: Street smth
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: '12345'
        phone_number:
          type: string
          example: '12345'
      xml:
        name: user
  securitySchemes:
    api_key:
      type: apiKey
      name: api_key
      in: header